- Refactoring vario per riordinare - grafica evaluate

+++ src/web/webapp.py
@@ -1,107 +1,26 @@
 from flask import Flask, render_template, request, Markup
 from werkzeug import secure_filename
-from src import mlp_main
+from src import mlp_wrapper
 import os
 
 app = Flask(__name__)
 
 root = "../../"
 
-train_new_network_html = """
-<div class="panel panel-primary">
-    <div class="panel-heading">
-        New custom Neural Net
-    </div>
-    <div class="panel-body">
-        <form action="http://localhost:5000/new_neural_net" method="POST" enctype="multipart/form-data">
-            <label> Neural network name </label> <input class="form-control" name="nn_filename" placeholder="nome.dat" />
-            <label> Hidden Layers</label> <br> values separated by commas!<input class="form-control" name="nn_hidden_layers" placeholder="64, 16" />
-            <br>
-            <label> Compute symmetry features <input type="checkbox" value="True" name="symmetry"> </label>
-            <br>
-            <br>
-            <label> Healthy training set</label> <input type="file" name="healthy_training"/><br>
-            <label> Stroke training set</label> <input type="file" name="stroke_training"/><br>
-            <label> No. of training epochs </label> <input class="form-control" name="epochs" placeholder="300" />
-
-            <input class="btn btn-default" type="submit"/><br></form>
-
-    </div>
-</div>
-"""
-
-upload_sample_html = """
-<div class="panel panel-primary">
-    <div class="panel-heading">
-        Upload a .dat file, choose the neural network and classify
-    </div>
-    <div class="panel-body">
-
-        <form action="http://localhost:5000/sample_classifier" method="POST" enctype="multipart/form-data">
-            <div>
-                    Choose an existing neural network.
-                <select class="form-control" name="nn_name">
-                    {}
-                </select>
-            </div>
-            <br><br>
-            <div>
-                Choose a <i>.dat</i> file to classify.
-                <input type="file" name="file"/><br>
-            </div>
-                <input class="btn btn-default" type="submit"/><br></form>
-    </div>
-</div>
-"""
-
-classification_result_html = """
-<div class="panel panel-default">
-    <div class="panel-heading">
-        Using neural network: <i>{}</i>
-    </div>
-    <div class="panel-body">
-        <div class="alert alert-info"> <i>{}</i> </div> has been classified as <b><font color="{}">{}</font></b>.
-
-    </div>
-
-"""
-
-evaluate_existing_nn_html = """
-<div class="panel panel-primary">
-    <div class="panel-heading">
-        Evaluate existing Neural Net
-    </div>
-    <div class="panel-body">
-        <form action="http://localhost:5000/evaluate_existing" method="POST" enctype="multipart/form-data">
-            <div>
-            Choose an existing neural network.
-            <select class="form-control" name="nn_name">
-                    {}
-             </select>
-            </div>
-            <br>
-            <br>
-            <label> Healthy test set</label> <input type="file" name="healthy_training"/><br>
-            <label> Stroke test set</label> <input type="file" name="stroke_training"/><br>
-            <input class="btn btn-default" type="submit"/><br></form>
-
-    </div>
-</div>
-"""
-
 
 @app.route('/')
 def index():
-    html = """
-    <object data="static/brain.svg" type="image/svg+xml">
+    img = """
+    <object data="static/brain.svg" type="image/svg+xml" height="80ems">
     </object>
     """
-    return render_template('index.html', page_inner=Markup(html))
+    return render_template('index.html', page_inner=Markup(img))
 
 
 @app.route('/train')
 def train():
-    return render_template('index.html', page_inner=Markup(train_new_network_html), train=True)
+    content = Markup(render_template('train_new_network.html'))
+    return render_template('index.html', page_inner=content, train=True)
 
 
 @app.route('/new_neural_net', methods=['GET', 'POST'])
@@ -120,9 +39,12 @@ def new_nn():
         stroke_training_file.save(s_tr_filepath)
         epochs = int(request.form['epochs'])
 
-        # TODO qualche bel controllino qui no eh?
-        mlp_main.train_new(root, nn_filename, h_tr_filepath,
-                           s_tr_filepath, epochs, symmetry, *hidden_layers)
+        if mlp_wrapper.check_exists_nn(root, nn_filename):
+            content = Markup(render_template('nn_exists_error.html', nn_name=nn_filename))
+            return render_template('index.html', page_inner=content)
+
+        mlp_wrapper.train_new(root, nn_filename, h_tr_filepath,
+                              s_tr_filepath, epochs, symmetry, *hidden_layers)
 
         os.remove(h_tr_filepath)
         os.remove(s_tr_filepath)
@@ -135,8 +57,9 @@ def evaluate():
     select_html = ""
     for nn in get_saved_nns():
         select_html += "<option>{}</option>".format(nn)
-    return render_template('index.html', page_inner=Markup(evaluate_existing_nn_html.format(select_html)),
-                           evaluate=True)
+
+    content = Markup(render_template('evaluate_existing_nn.html', select=Markup(select_html)))
+    return render_template('index.html', page_inner=content, evaluate=True)
 
 
 @app.route('/evaluate_existing', methods=['GET', 'POST'])
@@ -144,24 +67,25 @@ def evaluate_existing():
     if request.method == 'POST':
         nn_filename = request.form['nn_name']
 
-        healthy_training_file = request.files['healthy_training']
-        h_tr_filepath = "{}tmp/{}".format(root, healthy_training_file.filename)
-        healthy_training_file.save(h_tr_filepath)
+        healthy_test_file = request.files['healthy_test']
+        h_test_filepath = "{}tmp/{}".format(root, healthy_test_file.filename)
+        healthy_test_file.save(h_test_filepath)
 
-        stroke_training_file = request.files['stroke_training']
-        s_tr_filepath = "{}tmp/{}".format(root, stroke_training_file.filename)
-        stroke_training_file.save(s_tr_filepath)
+        stroke_test_file = request.files['stroke_test']
+        s_test_filepath = "{}tmp/{}".format(root, stroke_test_file.filename)
+        stroke_test_file.save(s_test_filepath)
 
-        auc, figure = mlp_main.test_existing(root, nn_filename, h_tr_filepath, s_tr_filepath)
+        auc, figure = mlp_wrapper.test_existing_nn(root, nn_filename, h_test_filepath, s_test_filepath)
 
         print("AUC: {}".format(auc))
 
-        html = "<img src={}>".format(figure)
+        content = Markup(render_template('evaluate_existing.html', nn_name=nn_filename, img=figure,
+                                         healthy_set=healthy_test_file.filename,
+                                         stroke_set=stroke_test_file.filename))
 
-        os.remove(h_tr_filepath)
-        os.remove(s_tr_filepath)
-        return render_template('index.html', page_inner=Markup(html),
-                               evaluate=True)
+        os.remove(h_test_filepath)
+        os.remove(s_test_filepath)
+        return render_template('index.html', page_inner=content, evaluate=True)
 
 
 @app.route('/classify')
@@ -170,7 +94,9 @@ def upload_file():
     for nn in get_saved_nns():
         select_html += "<option>{}</option>".format(nn)
 
-    return render_template('index.html', page_inner=Markup(upload_sample_html.format(select_html)), classify=True)
+    content = Markup(render_template('upload_sample.html', select=select_html))
+
+    return render_template('index.html', page_inner=content, classify=True)
 
 
 @app.route('/sample_classifier', methods=['GET', 'POST'])
@@ -183,7 +109,7 @@ def upload_file1():
 
         nn_filepath = root + "userspace/saved_nns/{}".format(nn_name)
 
-        if not mlp_main.classify(nn_filepath, f.filename):
+        if not mlp_wrapper.classify(nn_filepath, f.filename):
             res = "STROKE"
             col = "red"
         else:
@@ -191,15 +117,16 @@ def upload_file1():
             col = "green"
         os.remove(f.filename)
 
-        return render_template('index.html',
-                               page_inner=Markup(classification_result_html.format(nn_name, f.filename, col, res)))
+        content = Markup(
+            render_template('classification_result.html', nn_name=nn_name, sample_name=f.filename, color=col,
+                            result=res))
+
+        return render_template('index.html', page_inner=content)
 
 
 def get_saved_nns():
     return [name for name in os.listdir(root + "userspace/saved_nns/")
             if os.path.isdir(os.path.join(root + "userspace/saved_nns/", name))]
-    # dirs = [x[0] for x in os.walk(root + "userspace/saved_nns/")]
-    # print(dirs[1::])
 
 
 if __name__ == '__main__':
